feat: Convert all Integer data types to Long
- Changed all occurrences of Integer to Long to handle larger values.
- Updated method signatures and variable declarations accordingly.

+++ src/main/java/com/unitalk/common/model/entity/Employee.java
@@ -15,9 +15,9 @@ import java.sql.Timestamp;
 public class Employee {
 
     @Id
-    private Integer empId;
+    private Long empId;
 
-    private Integer loginNo;
+    private Long loginNo;
 
     private String deptId;
